use webpage tools npm package
