+++ package.json
@@ -7,15 +7,16 @@
   "author": "@Abletech",
   "license": "GPL",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-preset-es2015": "^6.24.1",
-    "uglify-js": "^3.4.9",
-    "webpack": "^3.12.0"
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "babel-loader": "^8.0.0",
+    "uglify-js": "^3.0.27",
+    "webpack": "^3.5.1"
   },
   "dependencies": {
-    "core-js": "^2.6.1"
+    "@addressfinder/addressfinder-webpage-tools": "1.1.2",
+    "core-js": "^2.5.0"
   },
   "scripts": {
     "watch": "webpack -w --config webpack.config.js",
